+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# [0.10.0] - 2026-04-10
+
+* Image reading fix
+* System prompt fix
+* removed add-dirs
+
 # [0.9.1] - 2026-04-03
 
 * /context supported